Update RegexScanner.rb
Changed how items are grouped for tagging.  Previously if there were multiple matches against the same item, the item would be added to the group multiple times.  It would report things like "tagging N items with tag" where N would be counting the same item multiple times.  It may attempt to add the same tag to the same item multiple times.  Changed grouping structure from array of items to hash of items to make sure for a given tag we only ultimately store that item once in the group.